requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
Flask==3.1.0
Flask-Cors==5.0.0
Flask-PyMongo==2.3.0
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
packaging==24.2
pymongo==4.10.1
python-dotenv==1.0.1
waitress==3.0.2
Werkzeug==3.1.3
